Fix: add missing core-database dependencies and PdfiumCore pageIdx.toInt()

+++ pdf-engine/src/main/java/com/folio/pdfengine/PdfiumNativeDocumentSource.kt
@@ -66,7 +66,7 @@ internal class PdfiumNativeDocumentSource(
                     entries.add(
                         PdfDocumentSource.TocEntry(
                             title = bm.title,
-                            pageIndex = bm.pageIdx,
+                            pageIndex = bm.pageIdx.toInt(),
                             depth = depth
                         )
                     )
